fix: treat clang's bool-literal constant spelling as producer-specific
`dumper_clang_expr._initializer_value` normalizes a lone boolean literal
via plain `str(bool)`, spelling `"True"`/`"False"` instead of either
backend's real `true`/`false` source text. Mark this `Fact.unsupported()`
in `normalize_header_ast`'s constants loop, the same treatment already
given to clang's compound-initializer fingerprint and castxml's opaque
`FunctionType` tag -- an unchanged boolean constant no longer reports a
spurious hybrid-merge conflict purely from clang's own stringification.

The decimal-integer/character/float-literal normalization residual stays
a documented, accepted limitation (pinned by a regression test): no safe
structural signal in the value text alone distinguishes a normalized
spelling from a genuine value difference; closing it needs a shared
literal-grammar normalizer or a per-backend original-token fact, a
model-shape decision for a future slice.

Split `tests/test_semantic_normalizer.py` a second time once these new
tests pushed it past the AI-readiness gate's 1200-line new-test-file cap
-- the artifact-recognition-primitive tests (unresolved-type sentinel,
opaque `FunctionType`, clang expression fingerprint, bool-literal) move
to a new sibling file, `tests/test_semantic_normalizer_artifacts.py`,
mirroring the production-code split of `semantic_normalizer_artifacts.py`
out of `semantic_normalizer.py` for the identical reason.
